chore: update error message (#273)
We have noticed that where we throw an error and only use e as the
message, we may not display readable and helpful information in the Pepr
logs. Throwing the error with a specific message results in the Pepr
logs displaying helpful information about what went wrong. We should
check Pepr and KFC for additional instances of throwing errors using
just e and update the error messages.

Co-authored-by: Case Wylie <cmwylie19@defenseunicorns.com>