fix: remove use undici for fetch function (#470)
## Description

Changes the fetch function to use undici fetch. Updates the k8sCfg to
return undici RequestInit instead of node-fetch RequestInit.

**Note:** - Do not merge is present on this PR because we are waiting
for December to release this as it is a big change to the surface area
of KFC and also affects Pepr. This PR removes the remaining node-fetch
from Pepr

## Related Issue

Fixes #459

<!-- or -->

Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging

- [ ] Test, docs, adr added or updated as needed
- [ ] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>