chore: contributor guide for kfc (#510)
## Description

Creates a contributor guide for KFC explaining how to test, run the
project, and contribute. Since the CODE_OF_CONDUCT is a prereq for
contributing, I added it into this PR but it can be split into a
different PR if we deem necessary.

## Related Issue

Fixes #479 

<!-- or -->

Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>