chore: remove nock (#550)
## Description

We switched from node-fetch to undici around November. Undici does not
work with nock and has a mocking agent that comes with the library
therefor removing the need for nock. I just noticed that although we are
not using nock it is still listed as a dependency from the dependabot
update.

## Related Issue

Fixes #

<!-- or -->

Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>