fix: prefer tokens from kubernetes/client-node (#569)
## Description

This extends the getToken method to accept a string bearer token. The
bearer token is extracted from the symbol map returned from
`kubernetes/client-node`. If there is not a bearer it will try to read a
token from the well-defined service account location. If not, then there
will not be a token.

We do not need to worry about expiring tokens here because this method
is called each time a `K8s(..).[call]` is called.

This repo is reproducing the bug and showing how it works in the end. 

https://github.com/cmwylie19/generic-eks-test

## Related Issue

Fixes #566 

<!-- or -->

Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [ ] Test, docs, adr added or updated as needed
- [ ] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>