fix: avoid redundancy and add clarity to contributing (#670)
## Description

Avoid redundancy and add clarity to `CONTRIBUTING.md` so that users know
when it is okay to not create an issue and rephrase the second line
around commits being signed since it is in there twice.

The reason for this PR is to release since Undici bumped. 

## Related Issue

Fixes #

<!-- or -->

Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>