chore: remove ignore pattern and ignore failing unit tests so we can bump kubernetes/client-node (#684)
## Description

The version of `@kubernetes/client-node` that we are on has not had an
update in 7 months.

https://www.npmjs.com/package/@kubernetes/client-node?activeTab=versions

We need to update since there are no security patches or updates going
into this version and it is fundamental for KFC.

We plan on writing integration tests to cover the failing unit tests,
then finally rewriting the unit tests.

## Related Issue

Fixes #617 
Fixes #591 

<!-- or -->

Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>