fix: handle eslint exclusions on `e2e/`  when `format:check` runs (#693)
## Description

#691 lints E2Es, but the configuration causes some issues when `npm run
format:check` or during lint with the commit hooks. We'll see an error
along the lines of:

```text
  0:0  error  Parsing error: "parserOptions.project" has been provided for @typescript-eslint/parser.
The file was not found in any of the provided project(s): e2e/webapp-v1alpha1.ts
```

Because we store `e2e` outside of `src`, our `tsconfig.json` file
considers `e2e` tests as outside of scope. What we can do is add a
tsconfig for the `e2e` tests and update the eslint config accordingly.

## Related Issue

Relates to #691 

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed