chore: explicitly include matrix in tsconfig (#731)
## Description

When shifting from CJS to ESM the `matrix.js` file broke. This updates
the `tsconfig.json` in `e2e/` to include that file and updates the file
to be compliant to the rules in TypeScript. It also updates the workflow
to call `matrix` from TypeScript

This PR is going to fail the e2e tests because it needs #728  to pass

## Related Issue

Fixes #730 

<!-- or -->

Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>