chore: e2e with vite (#728)
## Description

In upgrading `@kubernetes/client-node`, which is now an ESM module, we
encountered many issues with Jest. This PR switches to using Vite for
end-to-end tests and is currently a WIP. I’d like to resolve the final
e2e test before merging to ensure a high level of confidence. We also
need to release this soon after our next Pepr release in case any issues
arise, so we have time to address them.


Also: We were not linting our e2e tests with prettier like the rest of
the project and in deep debugging the code became hard to read so I have
removed the e2e's from the prettier ignore

needs #731 

## Related Issue

Fixes #722
Fixes #724 
Fixes #727
Fixes #723

<!-- or -->

Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>