feat: finalize in kfc (#825)
## Description

This PR adds the ability to add or remove finalizers from Kubernetes
resources. Due to limitations in Kubernetes around JSON Patch and
requiring a patch file Server Side Apply was the best option to ensure
we are able to both and remove without needing a patch file.

Due to the nature of Server Side Apply we first fetch the resource and
then Apply the new state of the resource with the updated finalizers

## Related Issue

Fixes #807 

<!-- or -->

Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>
Co-authored-by: Sam Mayer <samayer12@gmail.com>