fix: extend v1objectmetadata when crd defines it (#947)
## Description

UDS Core defines a ClusterConfig CRD with metadata defined. Normally it
is not important for the generated types to define metadata because
Kubernetes will inject it into the CR (instance) at runtime. In the case
of ClusterConfig, they are restricting the name.

## Related Issue

Fixes #939 

<!-- or -->

Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>