fix: standard two week release (#965)
## Description

This PR is for a standard two week release of dependency and security
updates. Also, while releasing
[lula](https://github.com/defenseunicorns/lula/pull/366) we realized
that for trusted publishing to work we need to be on `npm 11.5.1+` which
is why this PR bumps the node version.

## Related Issue

Fixes #

<!-- or -->

Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [ ] Test, docs, adr added or updated as needed
- [ ] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>